chore(web): grid type refactor
